Framework in place for built-for-you static content
