+++ frontend-react/src/components/Markdown/MarkdownDirectory.ts
@@ -1,38 +1,22 @@
+/* Used to instantiate a set of static pages, like BuiltForYouIndex
+ * or HowItWorks */
+import * as module from "module";
+
 export interface MarkdownPageProps {
     directories: MarkdownDirectory[];
 }
 
-interface MarkdownDirectoryInit {
-    title: string;
-    slug: string;
-    root: string;
-    files: string[];
-}
-
+/* Used to create objects that hold pointers to markdown directories and the
+ * info needed to query them. This is because we cannot access the filesystem
+ * at runtime */
 export class MarkdownDirectory {
     title: string;
     slug: string;
-    root: string;
-    files: string[];
+    files: module[];
 
-    constructor({ title, slug, root, files }: MarkdownDirectoryInit) {
+    constructor(title: string, slug: string, files: module[]) {
         this.title = title;
         this.slug = slug;
-        this.root = root;
         this.files = files;
     }
-
-    getUrl(item: string) {
-        if (!this.files.includes(item)) return undefined;
-        return `${this.root}/${item}`;
-    }
 }
-
-export const MadeForYouDirectories = [
-    new MarkdownDirectory({
-        title: "March 2022",
-        slug: "/built-for-you/march-2022",
-        root: "/content/built-for-you/2022-march",
-        files: ["update-2022-03-01.md, update-2022-03-15.md"],
-    }),
-];
