+++ frontend-react/src/content/built-for-you/2022-june.md
@@ -1,0 +1,3 @@
+# June 2022
+
+More test content, baby!
