+++ frontend-react/src/react-app-env.d.ts
@@ -1,1 +1,3 @@
 /// <reference types="react-scripts" />
+
+declare module "*.md";
